Move helper functions into src/helpers
